Rename StartupPlugin to StartupAction
